feat: refactor RedeployOfficerModal and streamline MainContractSection for improved contract redeployment flow
